restructured project and removed WordImpl, SentenceImpl and TextImpl
