+++ src/main/java/stanfordnlp/TextController.java
@@ -1,17 +1,15 @@
 package stanfordnlp;
 
+import edu.kit.kastel.mcse.ardoco.core.api.data.text.Text;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import stanfordnlp.corenlp.CoreNLPProvider;
-import stanfordnlp.corenlp.api.Text;
-import stanfordnlp.dto.TextDTO;
+import stanfordnlp.dtoconverter.ObjectToDtoConverter;
+import stanfordnlp.dtoconverter.dto.TextDTO;
 
 import java.io.ByteArrayInputStream;
-import java.io.FileInputStream;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.Properties;
 
 @RestController
 public class TextController {
@@ -25,7 +23,7 @@ public class TextController {
 
     private Text processText(InputStream text) {
         CoreNLPProvider nlpProvider = new CoreNLPProvider();
-        return nlpProvider.getAnnotatedText(text);
+        return nlpProvider.processText(text);
     }
 
     private TextDTO convertToDto(Text text) {
