--- src/main/java/stanfordnlp/corenlp/api/Phrase.java
@@ -1,24 +1,0 @@
-/* Licensed under MIT 2022-2023. */
-package stanfordnlp.corenlp.api;
-
-import org.eclipse.collections.api.list.ImmutableList;
-import org.eclipse.collections.api.map.ImmutableMap;
-import stanfordnlp.dto.PhraseType;
-
-public interface Phrase {
-    int getSentenceNo();
-
-    String getText();
-
-    PhraseType getPhraseType();
-
-    ImmutableList<Word> getContainedWords();
-
-    ImmutableList<Phrase> getSubPhrases();
-
-    boolean isSuperPhraseOf(Phrase other);
-
-    boolean isSubPhraseOf(Phrase other);
-
-    ImmutableMap<Word, Integer> getPhraseVector();
-}
