--- src/main/java/stanfordnlp/corenlp/api/Sentence.java
@@ -1,54 +1,0 @@
-/* Licensed under MIT 2021-2023. */
-package stanfordnlp.corenlp.api;
-
-import org.eclipse.collections.api.list.ImmutableList;
-
-/**
- * Represents a sentence within the document.
- */
-public interface Sentence {
-
-    /**
-     * Returns the sentence number (starting at {@code 0}.
-     *
-     * @return the sentence number
-     */
-    int getSentenceNumber();
-
-    /**
-     * sets the phrases
-     * 
-     * @param phrases the phrases
-     */
-    void setPhrases(ImmutableList<Phrase> phrases);
-
-    /**
-     * Return the sentence number used for human readably output.
-     * Therefore, this method calculates the sentence number starting with {@code 1}.
-     *
-     * @return the sentence number starting at one
-     */
-    default int getSentenceNumberForOutput() {
-        return getSentenceNumber() + 1;
-    }
-
-    /**
-     * Returns the words contained by this sentence
-     *
-     * @return the words contained by this sentence
-     */
-    ImmutableList<Word> getWords();
-
-    /**
-     * Returns the text of this sentence
-     *
-     * @return the text of this sentence
-     */
-    String getText();
-
-    default boolean isEqualTo(Sentence other) {
-        return other != null && this.getSentenceNumber() == other.getSentenceNumber() && other.getText().equals(this.getText());
-    }
-
-    ImmutableList<Phrase> getPhrases();
-}
